requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==41.0.3
Django==4.2.4
djangorestframework==3.14.0
passlib==1.7.4
Pillow==10.0.0
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.4
pytz==2023.3
simplejwt==2.0.1
sqlparse==0.4.4
typing==3.7.4.3
tzdata==2023.3
